requirements.txt
fastapi
uvicorn
websockets
google-genai
python-dotenv
jinja2
